requirements.txt
astroid==2.11.7
autopep8==1.6.0
click==8.1.3
dill==0.3.5.1
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.0
pylint==2.14.5
python-dotenv==0.20.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.2
Werkzeug==2.2.1
wrapt==1.14.1
